Move replaceFullSizeImage into the extending component, from helpers/
This is only used on one place,
so it is probably better in that class.
