+++ assets/src/stories-editor/helpers/index.js
@@ -1391,27 +1391,3 @@ export const getCallToActionBlock = ( pageClientId ) => {
 	const innerBlocks = getBlocksByClientId( getBlockOrder( pageClientId ) );
 	return innerBlocks.find( ( { name } ) => name === 'amp/amp-story-cta' );
 };
-
-/**
- * Replaces the 'full' image size with a custom size, which has a limited height.
- *
- * This prevents the user from selecting an image in the Image block that's too big.
- * For example, 3200 x 4000.
- */
-export const replaceFullSizeImage = () => {
-	const blockEditorStore = 'core/block-editor';
-	const ampStorySizeSlug = 'amp_story_page';
-	const initialSizes = select( blockEditorStore ).getSettings( 'imageSizes' ).imageSizes;
-	const sizesWithoutFullSize = initialSizes.filter( ( size ) => 'full' !== size.slug );
-
-	// If the AMP Story slug isn't present, add it.
-	if ( ! sizesWithoutFullSize.filter( ( size ) => ampStorySizeSlug === size.slug ).length ) {
-		sizesWithoutFullSize.push(
-			{
-				slug: ampStorySizeSlug,
-				name: __( 'AMP Story Max Size', 'amp' ),
-			}
-		);
-		dispatch( blockEditorStore ).updateSettings( { imageSizes: sizesWithoutFullSize } );
-	}
-}
